CLOUD-50108 Fix transactions in cloudbreak
