+++ core/src/main/java/com/sequenceiq/cloudbreak/controller/advice/UserContextControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.controller;
+package com.sequenceiq.cloudbreak.controller.advice;
 
 import javax.servlet.http.HttpServletRequest;
