+++ core/src/main/java/com/sequenceiq/cloudbreak/service/hostgroup/HostGroupService.java
@@ -3,11 +3,13 @@ package com.sequenceiq.cloudbreak.service.hostgroup;
 import javax.inject.Inject;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.sequenceiq.cloudbreak.domain.HostGroup;
 import com.sequenceiq.cloudbreak.repository.HostGroupRepository;
 
 @Service
+@Transactional
 public class HostGroupService {
 
     @Inject
